pyproject.toml
[project]
name = "wellnesstrack"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cloudinary>=1.44.1",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "gunicorn>=23.0.0",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-anthropic>=0.3.20",
    "langchain-core>=0.3.76",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "opencv-python-headless>=4.12.0.88",
    "pydantic>=2.11.7",
    "python-doctr>=1.0.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
